extract: toggleOtherRepository in otherRepositories
toggleOtherRepository has been extracted into its
own function and composed in a hook.
