+++ src/features/createAndEditTemplate/otherRepositories/core/ports.ts
@@ -1,0 +1,4 @@
+import { OtherUUID } from 'features/createAndEditTemplate/shared/types/types';
+
+// ports input
+export type ToggleSelectedOtherRepository = (uuid: OtherUUID) => void;
